in response to PR comments
