+++ subscription-engine/kafka-ws/examples/WSClient/Client.js
@@ -1,13 +1,9 @@
 /* eslint-disable no-console */
 // Example WS client
 const fs = require('fs');
-
 const WebSocket = require('ws');
-
 const superagent = require('superagent');
-
 const jwt = require('jsonwebtoken');
-
 const { promisify } = require('util');
 
 const jwtSignAsync = promisify(jwt.sign).bind(jwt);
